+++ src/meico/midi/legacy/MidiOld.java
@@ -1,8 +1,11 @@
-package meico.midi;
+package meico.midi.legacy;
 
 /**
  * Created by Axel Berndt.
  */
+
+import meico.midi.MidiFileReader;
+import meico.midi.MidiFileWriter;
 
 import javax.sound.midi.*;
 import java.io.File;